feat: AppHeader Component
